Revert "Merge pull request #75 from ProgrammingLife3/refactor/resizeadapter"
This reverts commit 87ed79ce8d27d8ab5c31992f36417e5ff3a238ca, reversing
changes made to d04d32880ddde959e4c1684ad9b83fcc5dbe79a8.

Conflicts:
	src/main/java/tudelft/ti2806/pl3/Application.java
	src/main/java/tudelft/ti2806/pl3/controls/ResizeAdapter.java
